corrections to Package Slider

+++ app/src/main/java/com/skylightapp/Classes/Profile.java
@@ -30,7 +30,6 @@ import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.Locale;
-import java.util.Objects;
 
 import static com.skylightapp.Classes.Customer.CUSTOMER_ID;
 import static com.skylightapp.Classes.Customer.CUSTOMER_TABLE;
@@ -1366,7 +1365,7 @@ public class Profile extends User implements Parcelable, Serializable, BaseColum
 
     }
     @Ignore
-    public void addSavings(int  profileId, int  customerId, int id, double amount, double numberOfDay, double total, int daysRemaining, double amountRemaining, String date, String status) {
+    public void addSavings(int  profileId, int  customerId, int id, double amount, int numberOfDay, double total, int daysRemaining, double amountRemaining, String date, String status) {
         dailyReports= new ArrayList<>();
         String savingsCount = "C" + (dailyReports.size() + 1);
         CustomerDailyReport customerDailyReport = new CustomerDailyReport(profileId,customerId, id, amount,numberOfDay,total,daysRemaining,amountRemaining,date,status);
